Unifies the create and edit project modals into a single component and adds it to the project list page
