+++ app/page.tsx
@@ -1,169 +1,19 @@
 'use client'
 
 import { useEffect, useState } from "react";
-import { Container } from "@/components/Container";
 import { DataGrid, GridColDef } from "@mui/x-data-grid";
 import { checkDeviceType } from "@/hooks/checkDeviceType";
-import {
-  Button,
-  FormControl,
-  IconButton,
-  Modal,
-  TextField,
-} from "@mui/material";
-import { Add, Check, Close } from "@mui/icons-material";
+import { Button, TextField } from "@mui/material";
+import { Add } from "@mui/icons-material";
 import Link from "next/link";
 import { NewProject, Project } from "@/types/types";
 import { useProjectListContext } from "@/context/ProjectListContext";
-        
-import './style.scss'
 import { observer } from "mobx-react-lite";
 import ModelStore from "./stores/modelStore";
+import { Container } from "@/components/Container";
+import ProjectModal from "@/components/Modals/ProjectModal";
 
-// TODO: Create a new modal component?
-const CreateProjectModal: React.FC<{
-  open: boolean
-  onClose: () => void
-  onConfirm: (newProject: NewProject) => void
-}> = ({ open, onConfirm, onClose }) => {
-  const [newProjectData, setNewProjectData] = useState<NewProject>({
-    name: '',
-    homeownerName: '',
-    homeownerPhone: '',
-    homeownerEmail: '',
-    homeownerAddress: '',
-  })
-
-  const handleDataChange = (fieldName: string, value: string) => {
-    setNewProjectData((prevData) => ({
-      ...prevData,
-      [fieldName]: value,
-    }))
-  }
-
-  const resetState = () => {
-    setNewProjectData({
-      name: '',
-      homeownerName: '',
-      homeownerPhone: '',
-      homeownerEmail: '',
-      homeownerAddress: '',
-    })
-  }
-
-  const isSaveButtonEnabled = Object.values(newProjectData).every(Boolean);
-
-  return (
-    <Modal
-      open={open}
-      className="createModal"
-      onClose={() => {
-        onClose()
-        resetState()
-      }}
-      aria-labelledby="new-project-modal"
-      aria-describedby="add-project-modal"
-    >
-      <div className="createModal__content">
-        <div className="createModal__header">
-          <p>New Project</p>
-          <IconButton
-            sx={{
-              borderRadius: '4px',
-              border: '1px solid #2196F3',
-              color: '#2196F3',
-              padding: '4px 10px',
-            }}
-            onClick={onClose}
-            aria-label="close">
-            <Close />
-          </IconButton>
-        </div>
-        <form className="createModal__form">
-          <FormControl>
-            <TextField
-              onChange={({ target }) =>
-                handleDataChange('homeownerName', target.value)
-              }
-              fullWidth
-              id="outlined-basic"
-              label="Client Name"
-              variant="outlined"
-              value={newProjectData.homeownerName}
-              required
-              placeholder="Client Name"
-            />
-          </FormControl>
-          <FormControl>
-            <TextField
-              onChange={({ target }) => handleDataChange('name', target.value)}
-              id="outlined-basic"
-              label="Project Name"
-              variant="outlined"
-              value={newProjectData.name}
-              required
-              placeholder="Project Name"
-            />
-          </FormControl>
-          <FormControl>
-            <TextField
-              onChange={({ target }) =>
-                handleDataChange('homeownerAddress', target.value)
-              }
-              id="outlined-basic"
-              label="Project Address"
-              variant="outlined"
-              value={newProjectData.homeownerAddress}
-              required
-              placeholder="Project Address"
-            />
-          </FormControl>
-          <FormControl>
-            <TextField
-              onChange={({ target }) =>
-                handleDataChange('homeownerEmail', target.value)
-              }
-              id="outlined-basic"
-              label="Client Email Address"
-              variant="outlined"
-              value={newProjectData.homeownerEmail}
-              required
-              placeholder="Client Email Address"
-            />
-          </FormControl>
-          <FormControl>
-            <TextField
-              onChange={({ target }) =>
-                handleDataChange('homeownerPhone', target.value)
-              }
-              id="outlined-basic"
-              label="Client Phone Number"
-              variant="outlined"
-              value={newProjectData.homeownerPhone}
-              required
-              placeholder="Client Phone Number"
-            />
-          </FormControl>
-        </form>
-        <div className="createModal__footer">
-          <Button
-            variant="contained"
-            startIcon={<Check />}
-            onClick={() => onConfirm(newProjectData)}
-            disabled={!isSaveButtonEnabled}
-            size="small"
-            sx={{
-              marginLeft: 'auto',
-            }}
-            color="primary"
-          >
-            Save
-          </Button>
-        </div>
-      </div>
-    </Modal>
-  )
-}
+import './style.scss'
 
 const  Home = observer(() => {
   const {
@@ -252,7 +102,7 @@ const  Home = observer(() => {
 
   return (
     <main>
-      <CreateProjectModal
+      <ProjectModal
         open={openModal}
         onConfirm={(newProject: NewProject) => handleCreate(newProject)}
         onClose={() => setOpenModal(false)}
